requirements.txt
pytest~=6.1.0
setuptools~=57.4.0
numpy~=1.19.2
pyserial~=3.5